Refactoring street_network in kraken
